Updated freeTubeUpdate.sh

+++ shl/freeTubeUpdate.sh
@@ -9,7 +9,10 @@ gitHubUser=FreeTubeApp
 gitHubRepo=FreeTube
 gitHubAPIRepoURL=$gitHubAPIURL/repos/$gitHubUser/$gitHubRepo
 freeTubeLatestGitHubReleaseURL=$(curl -s $gitHubAPIRepoURL/releases | jq -r '.[0].assets[] | select(.name | contains("amd64.deb")) | .browser_download_url')
-freeTubeLatestGitHubReleaseName=$(basename $freeTubeLatestGitHubReleaseURL)
-freeTubeLatestGitHubReleaseTag=$(curl -s $gitHubAPIRepoURL/tags | jq -r '.[0].name')
-wget -O $freeTubeLatestGitHubReleaseName "$freeTubeLatestGitHubReleaseURL"
-sudo gdebi -n $freeTubeLatestGitHubReleaseName && rm -v $freeTubeLatestGitHubReleaseName
+freeTubeLatestGitHubReleaseTag=$(curl -s $gitHubAPIRepoURL/tags | jq -r '.[0].name' | sed 's/v//;s/-beta//')
+freeTubeInstalledVersion=$(dpkg-query --showformat='${Version}' -W freetube)
+if [ "$freeTubeLatestGitHubReleaseTag" != "$freeTubeInstalledVersion" ];then
+	freeTubeLatestGitHubReleaseName=$(basename $freeTubeLatestGitHubReleaseURL)
+	wget -O $freeTubeLatestGitHubReleaseName "$freeTubeLatestGitHubReleaseURL"
+	sudo gdebi -n $freeTubeLatestGitHubReleaseName && rm -v $freeTubeLatestGitHubReleaseName
+fi
